add input formControl
