+++ src/app/app.module.ts
@@ -3,22 +3,20 @@ import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { InputClearableExampleComponent } from './input-clearable-example/input-clearable-example.component';
-import {MatInputModule} from '@angular/material/input';
 import { FormsModule } from '@angular/forms';
-import {MatIconModule} from '@angular/material/icon';
+import { MatIconModule } from '@angular/material/icon';
+import { CalculationsModule } from './calculations/calculations.module'
 
 @NgModule({
   declarations: [
-    AppComponent,
-    InputClearableExampleComponent
+    AppComponent
   ],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
-    MatInputModule,
     FormsModule,
-    MatIconModule
+    MatIconModule,
+    CalculationsModule
   ],
   providers: [],
   bootstrap: [AppComponent]
