+++ src/app/calculations/calculations.module.ts
@@ -1,0 +1,24 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
+import { InputClearableExampleComponent } from './components/input-clearable-example/input-clearable-example.component'
+import { MatInputModule } from '@angular/material/input';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+
+@NgModule({
+  declarations: [
+    InputClearableExampleComponent
+  ],
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    MatInputModule,
+    MatIconModule,
+    MatButtonModule
+  ],
+  exports:[
+    InputClearableExampleComponent
+  ]
+})
+export class CalculationsModule { }
