+++ src/app/calculations/components/input-clearable-example/input-clearable-example.component.css
