--- src/app/input-clearable-example/input-clearable-example.component.html
@@ -1,59 +1,0 @@
-<mat-form-field class="example-form-field">
-    <mat-label>Скорость инфузии, [мл/чac]</mat-label>
-    <input 
-    matInput type="number" 
-    min="0" 
-    [(ngModel)]="infusionRate"
-    >
-    <button 
-    mat-button 
-    *ngIf="value"
-    matSuffix
-    mat-icon-button 
-    aria-label="Clear"
-    (click)="value=''"
-    >
-    <mat-icon>close</mat-icon>
-    </button>
-</mat-form-field>
-<br>
-<mat-form-field class="example-form-field">
-    <mat-label>Время инфузии, [чac]</mat-label>
-    <input matInput 
-    type="number" 
-    min="0" 
-    [(ngModel)]="infusionTime"
-    >
-    <button mat-button 
-    *ngIf="value" 
-    matSuffix 
-    mat-icon-button 
-    aria-label="Clear" 
-    (click)="value=''"
-    >
-    <mat-icon>close</mat-icon>
-    </button>
-</mat-form-field>
-<br>
-<mat-form-field class="example-form-field">
-    <mat-label>Дозировка [мг]</mat-label>
-    <input matInput 
-    type="number" 
-    min="0" 
-    [(ngModel)]="dosage"
-    >
-    <button mat-button 
-    *ngIf="value" 
-    matSuffix 
-    mat-icon-button 
-    aria-label="Clear" 
-    (click)="value=''"
-    >
-    <mat-icon>close</mat-icon
-    >
-    </button>
-</mat-form-field>
-<br>
-<p>Общий объем инфузии: {{infusionRate*1 * infusionTime*1}} [мл]</p> 
-<p>Объем препарата: {{dosage*0.02}} [мл]</p>
-<p>Объем NaCl: {{(infusionRate*1 * infusionTime*1) - dosage*0.02}} [мл]</p>
